pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
dependencies = [
    'jira == 3.*',
    'python-dateutil == 2.*',
    'requests == 2.*',
    'SQLAlchemy == 2.*',
]
name = "reporting"
requires-python = ">=3.10"
version = "1.4.0"

[project.optional-dependencies]
dev = [
    "black",
    "factory_boy",
    "Faker",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
    "responses",
    "types-python-dateutil",
    "types-requests",
]

[project.scripts]
reporting = "reporting.cli:main"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = ["factory", "responses"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra"
minversion = "8.0"
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true
